Add files via upload
This script is designed to extract Landsat 8 Surface Reflectance NDVI averages from field boundaries and export the values as a table.  The date of the image is found in the header of the table and the NDVI averages as as the cells in the column corresponding to the date.  Here CLUID is used as the unique identifier for each field. 

This is intended for use in Google Earth Engine and can be pasted into the Code Editor. 

Created by Alison Thieme and John Fitz as part of the NASA DEVELOP Chesapeake Bay Ag project at NASA Goddard Space Flight Center.
Team Members: John Fitz, Sean McCartney, Perry Oddo, Alison Thieme, Sunita Yadav (Team Leader)
Last updated on April 10, 2018 by Alison Thieme. alisonthieme@gmail.com